fix performance problem in bootstrapping
So EscapeAnalysis.jl has a serious performance problem, which appears
only in bootstrapping.

I made some inspections using the following `print` debug:
> within `find_escapes!`
```julia
if length(argtypes) > 0
    Core.println(argtypes[1], " ", length(argtypes), " ", nstmts)
end
```
... and it seems like the performance of `find_escapes!` will be really
awful when there is a large number of statements, say, `nstmts > 100`.

The worse case I confirmed so far happens when analyzing
`construct_ssa!(::CodeInfo, ::IRCode, ::DomTree, ::Any, ::Vector{Any})`,
where I got the print `Core.Const(val=Core.Compiler.construct_ssa!) 182 
4262`,
and it took more than 100 sec on my machine (!!!).
Note that `find_escapes!` works on after-inlining IR, so we often such
many statements.

The interesting property is, this performance problem seems really
specific to bootstrapping. I confirmed that `find_escapes!` runs in
0.34 sec when I enabled it later using Revise.
```julia
julia> @time code_typed(Core.Compiler.construct_ssa!, 
(Core.Compiler.CodeInfo, Core.Compiler.IRCode,
Core.Compiler.DomTree, Any, Vector{Any}));
...
Core.Const(val=Core.Compiler.setindex!) 4 5
Core.Const(val=Core.Compiler.construct_ssa!) 182 4262
0.342626 seconds (723.68 k allocations: 177.828 MiB, 15.69% gc time, 
99.94% compilation time)
```
